pyproject.toml
[project]
name = "bmailer"
version = "0.1.0"
description = ""
authors = [{ name = "Duy Tan Bui", email = "duytanisme02@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "python-dotenv",
  "jinja2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
